+++ R/item_reliability.R
@@ -5,21 +5,25 @@
 #'    for tests or item-scales of questionnaires.
 #'
 #' @param x A matrix or a data frame.
+#' @param type Type of reliability estimate. Either `"alpha"` (default, see
+#' [`cronbachs_alpha()`]) or `"omega"` (see [`mcdonalds_omega()`]). Note that
+#' computing McDonald's Omega is more computationally intensive than Cronbach's
+#' Alpha.
 #' @param standardize Logical, if `TRUE`, the data frame's vectors will be
 #'   standardized. Recommended when the variables have different measures /
 #'   scales.
 #' @param digits Amount of digits for returned values.
 #'
 #' @return A data frame with the corrected item-total correlations (*item
-#'      discrimination*, column `item_discrimination`) and Cronbach's Alpha
-#'      (if item deleted, column `alpha_if_deleted`) for each item
-#'      of the scale, or `NULL` if data frame had too less columns.
+#' discrimination*, column `item_discrimination`) and Cronbach's Alpha
+#' (if item deleted, column `alpha_if_deleted`) resp. McDonald's Omega for each
+#' item of the scale, or `NULL` if data frame had too few columns.
 #'
 #' @details
-#'
 #' This function calculates the item discriminations (corrected item-total
 #' correlations for each item of `x` with the remaining items) and the
-#' Cronbach's alpha for each item, if it was deleted from the scale. The
+#' Cronbach's alpha (when `type = "alpha"`) resp. McDonald's Omega (when
+#' `type = "omega"`) for each item, if it was deleted from the scale. The
 #' absolute value of the item discrimination indices should be above 0.2. An
 #' index between 0.2 and 0.4 is considered as "fair", while an index above 0.4
 #' (or below -0.4) is "good". The range of satisfactory values is from 0.4 to
@@ -28,17 +32,22 @@
 #' determine why a negative value was obtained (e.g. reversed answer categories
 #' regarding positive and negative poles).
 #'
-#' @examples
+#' @examplesIf requireNamespace("lavaan", quietly = TRUE)
 #' data(mtcars)
-#' x <- mtcars[, c("cyl", "gear", "carb", "hp")]
-#' item_reliability(x)
+#' item_reliability(mtcars[, c("cyl", "gear", "carb", "hp")])
+#'
+#' data(iris)
+#' item_reliability(iris[1:4], type = "omega")
 #' @export
-item_reliability <- function(x, standardize = FALSE, digits = 3) {
+item_reliability <- function(x, type = "alpha", standardize = FALSE, digits = 3) {
   # check param
   if (!is.matrix(x) && !is.data.frame(x)) {
     insight::format_alert("`x` needs to be a data frame or matrix.")
     return(NULL)
   }
+
+  # alpha or omega?
+  type <- match.arg(type, c("alpha", "omega"))
 
   # remove missings, so correlation works
   x <- stats::na.omit(x)
@@ -58,20 +67,33 @@ item_reliability <- function(x, standardize = FALSE, digits = 3) {
     # when items have different measures / scales
     if (standardize) x <- .std(x)
 
-    # calculate cronbach-if-deleted
-    cronbachDeleted <- vapply(seq_len(ncol(x)), function(i) cronbachs_alpha(x[, -i]), numeric(1L))
-
     # calculate corrected total-item correlation
     totalCorr <- vapply(seq_len(ncol(x)), function(i) {
       stats::cor(x[, i], rowSums(x[, -i]), use = "pairwise.complete.obs")
     }, numeric(1L))
 
-    ret.df <- data.frame(
-      term = df.names,
-      alpha_if_deleted = round(cronbachDeleted, digits),
-      item_discrimination = round(totalCorr, digits),
-      stringsAsFactors = FALSE
-    )
+    # reliability estimate alpha or omega?
+    if (type == "alpha") {
+      # calculate cronbach-if-deleted
+      cronbachDeleted <- vapply(seq_len(ncol(x)), function(i) cronbachs_alpha(x[, -i]), numeric(1L))
+
+      ret.df <- data.frame(
+        term = df.names,
+        alpha_if_deleted = round(cronbachDeleted, digits),
+        item_discrimination = round(totalCorr, digits),
+        stringsAsFactors = FALSE
+      )
+    } else {
+      # calculate omega-if-deleted
+      omegaDeleted <- vapply(seq_len(ncol(x)), function(i) mcdonalds_omega(x[, -i], ci = NULL), numeric(1L))
+
+      ret.df <- data.frame(
+        term = df.names,
+        omega_if_deleted = round(omegaDeleted, digits),
+        item_discrimination = round(totalCorr, digits),
+        stringsAsFactors = FALSE
+      )
+    }
   } else {
     insight::format_warning("Data frame needs at least three columns for reliability-test.")
   }